feat: TASK-04 — Float → Numeric(12,2) finansal kolonlar
Backend:
- 8 model: Float → Numeric(12,2), Mapped[float] → Mapped[Decimal]
  (auction, bid, listing, listing_offer, purchase, search_alert,
   user.max_budget, direct_sale/order Numeric(10,2)→(12,2))
- schemas/base.py: BaseSchema (Decimal→float JSON serializer)
- schemas/auction, direct_sale, listing: output alanları Decimal
- utils/price.py: to_price() yardımcısı
- database_clickhouse.py: Decimal(10,2) → Decimal(12,2)
- alembic zzzzw_float_to_numeric: 16 ALTER TABLE

Flutter:
- number_formatter.dart: forceDecimals parametresi (her zaman 2 hane)
- listing_detail_screen: toInt() fiyat kesme hatası düzeltildi
- create/edit listing + direct_sale_panel: allowDecimal: true
- direct_sale_panel fiyat gösterimleri: forceDecimals: true
- direct_sale_panel toplam hesabı: toStringAsFixed(2) precision fix
- seller_report_screen: _fmtBudget → TeqNumberFormatter
- pro_insights_screen + swipe_live_screen: forceDecimals: true
- auction_panel: as double? → (as num?)?.toDouble() güvenli cast

+++ backend/app/models/auction.py
@@ -1,6 +1,7 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
-from sqlalchemy import String, Integer, Float, DateTime, ForeignKey, Boolean, func
+from sqlalchemy import String, Integer, Numeric, DateTime, ForeignKey, Boolean, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -13,9 +14,9 @@ class Auction(Base):
     stream_id: Mapped[int] = mapped_column(ForeignKey("live_streams.id"), nullable=False, index=True)
     listing_id: Mapped[Optional[int]] = mapped_column(ForeignKey("listings.id", ondelete="SET NULL"), nullable=True, index=True)
     item_name: Mapped[str] = mapped_column(String(300), nullable=False)
-    start_price: Mapped[float] = mapped_column(Float, nullable=False)
-    buy_it_now_price: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
-    final_price: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
+    start_price: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
+    buy_it_now_price: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 2), nullable=True)
+    final_price: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 2), nullable=True)
     is_bought_it_now: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     winner_id: Mapped[Optional[int]] = mapped_column(ForeignKey("users.id"), nullable=True, index=True)
     winner_username: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
